Restructure artifact files model.
The new model is an artifact can have multiple "root" files associated
with it.  If we extract any files for display/indexing, we place those
into a separate table, artifact_file_members.  This removes the pain
of a recursive foreign key from the artifact_files table.  We do not
need the unlimited hierarchy expressed by the previous schema.