Bugfix: only autodelete referenced objects that have a non-null fk into referencing object.
This code exists for the case where we are processing object X's relations, and one
of those relations is being removed, and the related object has a non-nullable fk
reference into X.  Simply removing the related object from our list has the effect
of setting the non-nullable fk to None; so remove the related object first.

Anyway, we were not checking to see if the related object had an fk into X's table;
now we do.