Remove login match to existing Person; only match Users.
We don't want to match an existing Person, because that Person does
most likely match the incoming user (since the comparison is email
addr), *but* the user might not actually want to be associated with
every artifact/affiliation that the matched Person is associated
with (similar issue as taking artifact ownership after the fact;
we can do the same with Person relationships).

So if there is no User match, create both a new Person and User.