Fix additional bugs in /login.
In particular, since we don't lock tables in /login, we need to keep
user/person/session creation all in the same transaction, and if it
fails for possible duplicate, try again to validate the incoming info.
This strategy relies on tightening the unique constaint on Sesssions
from (user_id,sso_token) to (sso_token) only.  Short of a collision,
it is impossible for a token to map to multiple users; and a single
user should not have multiple sessions with the same token.