Refactor admin_status importer service handling.
We were overloading admin_status as something that the backend
determined by probing; and something that an admin could set
statically to disable an importer.  Now, the backend only attempts
to bring an importer to enabled admin_status if the importer is new;
otherwise, that bit is only flipped by an admin.