Split UserAffiliations off from Affiliations.
We were using Affiliations as part of profiles, and as part of
ArtifactAffiliations.  This led to the problem of "emergent" affiliations
in profiles, where an ArtifactAffiliation that related the User's
Person/Affiliation/Org to an Artifact (and maybe the Org was null) would
cause this Affiliation to show up in the User's profile.  This is not
desired sharing for profile affiliations, and is the cause of the null org
bug in https://github.com/ITI/searcch/issues/41, so we just halt the reuse of
Affiliation for profile stuff, and hope for the best down the road.  For
instance, this may cause us to merge Affiliation into ArtifactAffiliation
just to simplify things; not sure on that yet.

Fixes https://github.com/ITI/searcch/issues/41 .