Add a backend mail service configuration based on postfix.
Either you use it raw and allow it to connect to destination MTAs
directly; or you can specify the RELAYHOST (and optionally RELAYUSER,
RELAYPASS env vars to relay mail through another MTA).

In direct mode, we don't yet hook up postfix with certs, and that
will probably ding us a bit on spam score for public hosts.  This is
not straightforward since we're relying on caddy to automatically
maintain our certs and caddy can't export its live certs, nor can
we assume anything about the naming.  We'll solve that later.