Bugfix: actually reverse the ArtifactRelationship keys/relationships for Artifact.reverse_relationships.
Previously this was not required, so the contents of Artifact.reverse_relationship
were allowed to be mostly incorrect.  This fixes things so that a faked ArtifactRelationship
is created for each real relationship, with artifact/related_artifact fields swapped,
and artifact_group replacing related_artifact_group.  The only thing we don't change is the
name of the relation, since the backend doesn't know the inverse names yet.