Ensure background tasks only run in gunicorn arbiter; expose config bit for email invite interval.
This is a simple, stupid approach: start the scheduler in arbiter via the
on_starting hook (as we were already doing); but then *stop* the scheduler
in worker threads via the post_init hook (which is post-fork).  This is
an obviously impoverished and hacky approach, but I have never been able
to find a way to guarantee that something runs only in the arbiter, as the
on_starting hook allows.